refactor(server): move snapshot-adjacent services into package
